Split the configuration into     autoconf/configure.begin.in     autoconf/configure.end.in     autoconf/acinclude.m4     configure.in.in
and add a autogen.sh script to generate configure.in and then configure.
This script is automatically called by end.pl when needed

Added the needed genom macros for package dependencies (-P option)
Added the regen-user target
Added NEWARGS in regen(-user)? to add new arguments without editing genom.mk
Added BINEXPORT and set it to -export-dynamic in POSIX. It is needed
on recent libcs to allow dlopen on self