58. Don't initialize pocolibs timers if there is no periodic task.
This is done by calling h2InitGlob(0) instead of 100 in moduleInit.c when
every execution task has a period of 0.
This is mainly intended to work around a probable bug in gdb that is
unable to step the execution of a thread after a SIGALRM has been
received. It should also have a positive impact on modules performances,
but this is probably minor.

... I hope this change has no side effect?