Add -I$prefix/include to cflags for module clients.
In theory we should be able to get rid of the -I$prefix/include/module
option, since everyone should use <module/header.h> and not <header.h>
when including a header, unfortunatly this is not yet true.

And since pocolibs already pulls a -I$prefix/include option, this is
almost redundant, except for a few cases, like if pocolibs and genom are
installed with different prefixes.