Don't call aclocal in modules' autogen scripts.
First, aclocal is from automake and genom doesn't use automake. So better not
depend on automake.

Then, running aclocal might pull-up a libtool.m4 that do not correspond to the
ltmain.sh bundled with genom. This is what was actually breaking the
builds on recent ubuntu or debian systems. Furthermore, it's not really
possible to regenerate ltmain.sh with libtoolize because of the screwed up
AC_CONFIG_AUX_DIR variable.

As a consequence, and because we have to regenerate the modules' configure
script, we have to include and distribute libtool.m4 in the canvas so that
autoconf can find it. We also have to include pkg.m4 (from pkg-config) for the
same reason.

The drawback is that people that will need new m4 features in configure.ac.user
will have to explicitely include the required m4 files. If this happen to be
painful, this commit might be reverted ...