Make sure that include files are found from local dir before anything else
If a CPPFLAGS is inherited from the environment and it contains include
directives where some genom header files are already present, this might make
the compilation fail (or generate a wrong binary).
So make sure that local directories are always scanned first.

This fixes the "make replace" target of genom in robotpkg.