Fix several problems related to client disconnection in tclserv
If a client disconnects while he has pending requests (e.g. sent with
-ack), those requests must not be freed because:

. if the Tcl request id is reused and a reply corresponding to the old
  request arrives this will mess up the requests state.
. if the Tcl channel is reused by a new client, this client might
  receive spurious messages associated to the old requests, messing up
  the new client's state.
. if replies for the old requests arrive later, and no associated id is
  present in the state, the replies might never be read, filling up
  tclserv's mailbox.

So instead of deleting the requests, we assign them an empty client
channel and leak the requests id temporarily. If final replies arrive
later, the answer is not sent to anyone (thanks to the empty client
channel) and the request id is automatically freed at this point.

This is really amazing that tclserv could be used for 10 years now
with this kind of bug inside...