Fix posters_input attribute when a poster type is repeated in different fields
For each request or task, the posters_input is processed as a list. If
a type in the list had already been processed in another request, the
processing of the full list was interrupted instead of skipping to the
next element.

Fixes https://softs.laas.fr/bugzilla/show_bug.cgi?id=92