Fix (finally !) the "my tclserv is eating 100% cpu" bug
When `h2 end` was done while tclserv was running and tclserv had already
processed some requests (i.e the thread listening on the tclserv mailbox was
blocked in h2SemTake()), tclserv would start eating 100% cpu.

This was because tclserv assumed that h2SemTake() would return FALSE in case of
a timeout and ERROR in case of error. While this may have been true a very long
time ago, h2SemTake() nowadays always returns FALSE in both situations (and
sets errno to TIMEOUT in the case of a timeout).

tclserv now correctly deals with h2SemTake() returning an error after `h2 end`
and even sends a 'DIE' request to itself, so that it will shutdown gracefully.
Note: this will also terminate all clients connected to that server, which is
questionnable.