Fix a glitch for in-tree builds.
When building in-tree, the VPATH in $(top_srcdir)/lib introduced in 2ce30b0
finds the printScanUtil.lo libtool object (in lib) and don't rebuild it
locally. It is necessary to rewrite the genom binary linkage rule could to use
$^ instead of ${OBJS} so that the VPATH to printScanUtil.lo is propagated to
the link command. Out-of-tree builds did not have this issue.

This problem was introduced in 2ce30cb0. Hi Matthieu!