Add cs::lsmbox procedure in tclserv
cs::lsmbox takes no argument and return the list of available 'server'
mailboxes on the host. 'reply' mailboxes, as created by gcomLib, are filtered
out by checking if their name end in 'R'. This is not robust, but since only
mboxLib interface is visible at the h2dev level, there is no other way.

Note that to do something useful with this list, some other knowledge - for
instance a mapping between a program name and its mailbox name - has to be
injected at a higher level. The name of a server mailbox is not related to
anything else, and in particular not to the program that created it. It happens
that genom modules create their server mailbox after the name of the program,
but this is only a matter of genom convention and might break for other kind of
program using mboxLib.