Add 'rexec server args' procedure in genom tcl package
The procedure run args on server. args may be any string valid for the 'exec'
tcl command (which see). args is always run in the background, and the standard
channels of the program are those of the tclserv process. The procedure returns
the pid of the spawned process. If you need to catch the standard output, use
'server::reval server exec args' instead.

rexec is a trivial facility on top of the 'server::reval' procedure, so it
suffers from the same flaws. It is, in particular, a huge security hole for the
user that runs it. ** NEVER run tclserv in an untrusted environment **. You've
been warned ;)