Rewrite generation of the Requires: field in the .pc files
The new rule is to include all packages listed in the "requires" field of the
.gen file, and ignore dependencies listed in "codels_requires".
In addition, all packages mentionned in an "import from" directive must also be
listed in a require. A warning is printed otherwise.

For -oprs and -tclserv_client .pc files, the list of import_from is used to
find a match in the require field and output the -oprs or -tclserv_client
versions of the .pc files for genom modules only.
While here, added the main .pc file to the Requires of -oprs and
-tclserv_client pc files.