Revert part of 0d1bfd745a4e5b564b6d9652803b12b29b80a96a
Using sizeof(typename) is not that easy. Breaks badly with arrays or
pointers, and more subtily with struct typedefs if the struct name is
not the same as the typedef name.

In fact this is one more area that exposes the fundamental flaw in
Genom's types handling. Fortunatly we manage to hide it.