Check that module 'number' is positive and less than 32767 (15bits)
The module number (and thus module error numbers) is stored on the 16 high bits
of a 32 bits integer. Furthermore, module numbers must fit on 15 bits as
documented in h2errorLib.h:  M_id and err are encoded on 'signed short' (ie, <
2^15 = 32768)).

So check that module numbers fit on 15 bits in the parser. Otherwise strange
(untrackable) errors may happen.