Added support for having OpenPRS to talk to TclservClient.
Added the code to get the tclservclient functions assicated to each request.
It works for the demo-genom module together with the demo-transgen supervisor.

For now, -o assumes -x, but I will try to make it smarter and only link to the code when -x is used.

If people want to give it a spin... to test it more thorougly, let me know.