Properly handle the -x or not when generating openprs code (with -o).
If -x is not given, we do not generate code using tclserv
client. Tested with demo-genom and demo-transgen, both with cslib and
tclserv client. Seem to work as expected.

Most of the generated code is now in GenoM anyway, transgen is
becoming an empty shell, just making "one" call to a function defined
by genom which does all the work.