Add a reentrant version of a bunch of functions in printScanUtil.c
Make a reentrant version of indentStr, indentStr2, getIndexesStr and
getIndexesStr2. This is required when those functions are used from the
printXML* function, in the context of a multi-threaded picoweb, as noticed by
adegroot@.

Newly generated code will use the reentrant version, but old
generated code will still access the non-rentrant version which is kept for
compatibility.