Redo the instance name cache.
Caching the value for the 1st call of genomInstanceName() is not
good enough, when a C client calls several xxxClientInit() functions.

So cache the instance name per module name.

This still won't allow  a C client tries to use xxxClientInit() twice
for 2 instances of the same xxx module. you need to call csClientInit()
manually in this case.